% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timelines-set.R
\name{set_margins}
\alias{set_margins}
\title{Set margins for \code{timelines-chart}}
\usage{
set_margins(timelines, left = 90, right = 100, bottom = 26, top = 30)
}
\arguments{
\item{timelines}{A \code{timelines-chart} \code{htmlwidget} object.}

\item{left}{Chart's left margin, which contains the left-side group axis labels.}

\item{right}{Chart's right margin, which contains the right-side series axis labels.}

\item{bottom}{Chart's top margin, which contains the color legend.}

\item{top}{Chart's bottom margin, which contains the time axis labels.}
}
\description{
Set margins for \code{timelines-chart}
}
