% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timelines-set.R
\name{set_overview}
\alias{set_overview}
\title{Set overview options for \code{timelines-chart}}
\usage{
set_overview(timelines, enable = TRUE, domain = NULL)
}
\arguments{
\item{timelines}{A \code{timelines-chart} \code{htmlwidget} object.}

\item{enable}{Whether to show an interactive timeline overview below the chart.}

\item{domain}{The time extent shown in the overview section below the chart.}
}
\description{
Set overview options for \code{timelines-chart}
}
