% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timelines-set.R
\name{set_scale}
\alias{set_scale}
\title{Set scale for \code{timelines-chart}}
\usage{
set_scale(timevis, qualitative = FALSE, data_label = "", label = "")
}
\arguments{
\item{timevis}{A \code{timelines-chart} \code{htmlwidget} object.}

\item{qualitative}{Whether the segment data color values are categorical (false) or quantitative (true).}

\item{data_label}{Units of z data. Used in the tooltip descriptions.}

\item{label}{Color scale label. Only applicable to quantitative z scales.}
}
\description{
Set scale for \code{timelines-chart}
}
