% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timelines-set.R
\name{set_time}
\alias{set_time}
\title{Set time options for \code{timelines-chart}}
\usage{
set_time(timelines, use_utc = FALSE, format = "\%Y-\%m-\%d \%-I:\%M:\%S \%p")
}
\arguments{
\item{timelines}{A \code{timelines-chart} \code{htmlwidget} object.}

\item{use_utc}{Whether to display time in the local time zone (false) or in UTC (true).}

\item{format}{Time format to use in tooltips. See \href{https://github.com/d3/d3-time-format#locale_format}{d3-time-format} for available options.}
}
\description{
Set time options for \code{timelines-chart}
}
