% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timelines-set.R
\name{set_zoom}
\alias{set_zoom}
\title{Set zoom for \code{timelines-chart}}
\usage{
set_zoom(timelines, x = NULL, y = NULL)
}
\arguments{
\item{timelines}{A \code{timelines-chart} \code{htmlwidget} object.}

\item{x}{Chart's time (horizontal) zoom. A null value indicates a zoom reset to full extent.}

\item{y}{Chart's vertical zoom. The parameter should follow the syntax \code{c("<start row index>", "<end row index>")}.
A null value indicates a zoom reset to full extent.}
}
\description{
Set zoom for \code{timelines-chart}
}
