% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timelines-package.R
\docType{package}
\name{timelines-package}
\alias{timelines-package}
\title{An \code{htmlwidget} interface to the
timelines-chart javascript chart library}
\description{
This package allow you to use timelines-chart (\url{https://github.com/vasturiano/timelines-chart}),
a parallel timelines layout (swimlanes) for representing state of time-series over time.
}
\author{
Victor Perrier (@dreamRs_fr)
}
