% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timelines-proxy.R
\name{timelines_proxy}
\alias{timelines_proxy}
\title{Proxy for `timelines`}
\usage{
timelines_proxy(shinyId, session = shiny::getDefaultReactiveDomain())
}
\arguments{
\item{shinyId}{single-element character vector indicating the output ID of the
chart to modify (if invoked from a Shiny module, the namespace will be added
automatically).}

\item{session}{the Shiny session object to which the chart belongs; usually the
default value will suffice.}
}
\description{
Proxy to interact with an existing [timelines()] chart in Shiny.
}
