% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timelines-proxy.R
\name{update_data}
\alias{update_data}
\title{Update data for \code{timelines-chart} with proxy}
\usage{
update_data(timelines, data)
}
\arguments{
\item{timelines}{A \code{timelines-chart} \code{proxy} object.}

\item{data}{A \code{data.frame}, see \code{\link{timelines}} for details.}
}
\description{
Update data for \code{timelines-chart} with proxy
}
